+++ src/ast/import_visit.rs
@@ -140,6 +140,7 @@ impl ProvenanceValidator<'_> {
             PackageBodyElement::ViewUsage(n) => self.view_body(&n.value.body),
             PackageBodyElement::PerformUsage(n) => self.perform_body(&n.value.body),
             PackageBodyElement::ExtendedDefinition(n) => self.package_body(&n.value.body),
+            PackageBodyElement::KermlClassifier(_) => Ok(()),
             PackageBodyElement::Error(_)
             | PackageBodyElement::Unsupported(_)
             | PackageBodyElement::Doc(_)